requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==2.3.2
frozenlist==1.6.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.4.4
numpy==2.2.6
oauthlib==3.2.2
openai==1.82.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.2
pillow==11.2.1
priority==2.0.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
twilio==9.0.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0

zipp==3.19.0